Fix the RM_GetUsername crash when the user isn't set for the context.
Adds the necessary checks for the pointers to dereference prior to
dereferencing those to avoid crashes when the user is not set in
the redis module context.

This may happen both, accidentally and on purpose, depending on the
place of invocation of the function and how the redis module commands
or other functions were invoked by the redis server.